requirements.txt
websocket-client
websocket
django-cors-headers
